New test helper for snapshotting pg queries
