+++ posthog/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+# :TRICKY: Integrate syrupy with unittest test cases
+@pytest.fixture
+def unittest_snapshot(request, snapshot):
+    request.cls.snapshot = snapshot
